fixing loop size_t

+++ util.cpp
@@ -17,9 +17,9 @@ std::set<std::string> parseStringToWords(string rawWords)
 {
     std::set<std::string> list;
     string tempString = convToLower(rawWords);
-    for (int i = 0; i < tempString.size(); i++){
+    for (size_t i = 0; i < tempString.size(); i++){
         string temp = "";
-        for (int j = i; j < tempString.size(); j++){
+        for (size_t j = i; j < tempString.size(); j++){
             if (tempString[j] == ' ' || (ispunct(tempString[j]) && tempString[j] != '-')){
                 if (j == tempString.size() - 1 && tempString[j] != ' ' && !ispunct(tempString[j])){
                     temp += tempString[j];
